\name{cellchatrs_available}
\alias{cellchatrs_available}
\title{Is the Rust kernel loaded and callable?}
\description{
Is the Rust kernel loaded and callable?
}
\usage{
cellchatrs_available()
}
\arguments{
\item{...}{Arguments passed through unchanged. See the corresponding function in the \pkg{CellChat} package.}
}
\keyword{internal}
