\name{cellchatrs_threads}
\alias{cellchatrs_threads}
\title{Number of worker threads the Rust kernel is using}
\description{
Number of worker threads the Rust kernel is using.
}
\usage{
cellchatrs_threads()
}
\arguments{
\item{...}{Arguments passed through unchanged. See the corresponding function in the \pkg{CellChat} package.}
}
\keyword{internal}
